% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRAGadget.R
\name{QRAGadget}
\alias{QRAGadget}
\title{A Shiny Gadget for Interactive QRA Visualizations.}
\usage{
QRAGadget()
}
\description{
Upload raster data and easily create interactive QRA visualizations. Select
from numerous color palettes, basemaps, and different configurations.
}

